feat: Add option to choose whether to commit or just display message suggestions
Added a new option `-c` or `--commit` which can be used to select whether to create a commit or just display message suggestions. The default value is `true`, which means a commit will be created at the end of the program. If `false`, the program will only display message suggestions without creating a commit.