feat: Add option to append disclaimer to commit message
Added option `--disclaimer` (or `-d`) to append the message generated by `gpt-commit-messages` with a disclaimer at the end. The disclaimer informs that the message has been generated with `gpt-commit-messages` and provides a link to the GitHub repository.

Disclaimer: This message has been generated with gpt-commit-messages: https://github.com/Goddchen/gpt-commit-messages
Signed-off-by: Goddchen <goddchen@gmail.com>